check existing vis before run

+++ code/model/run_inference_show.py
@@ -6,52 +6,67 @@ import sys
 import cv2
 
 # Load a pretrained YOLO model
-model = YOLO("models/best21.pt")
+model = YOLO("runs/detect/train21/weights/best.pt")
+modelname = "best21_medium"
 
-vid = "/Users/jonas/Downloads/vid/NVR_Hien_EJDER7_2024-05-16_07.00.00_001000_001200.mp4"
-
+#vid = "/Users/jonas/Downloads/vid/NVR_Hien_EJDER7_2024-05-16_07.00.00_001000_001200.mp4"
+vids = list(Path("../../../../../../mnt/BSP_NAS2_work/eider_model/eider_testvids/").rglob("*.mp4"))
 
 # Run inference using the pretrained model and the inout video
 
+for vid in vids: 
 
-cap = cv2.VideoCapture(vid)
+    cap = cv2.VideoCapture(vid)
 
-writer = cv2.VideoWriter('out.mp4', cv2.VideoWriter_fourcc(*'mp4v'), 10.0, (640,360))
-
-while cap.isOpened():
-    # Read a frame from the video
+    # Read the first frame to get the frame size
     success, frame = cap.read()
+    if not success:
+        print(f"Failed to read the video: {vid}")
+        continue
 
-    if success:
-        # Run YOLOv8 tracking on the frame, persisting tracks between frames
-        results = model.predict(vid, 
-                    stream=True, 
-                    save = False,
-                    show = True)
+    # Get the frame size
+    frame_height, frame_width = frame.shape[:2]
 
-        # Get the boxes and track IDs
-        boxes = results[0].boxes.xywh.cpu()
+    name = "video_out/" + vid.stem + "_" + modelname + ".mp4"
     
-        # Visualize the results on the frame
-        annotated_frame = results[0].plot(line_width=1)
+    if Path(name).exists():
+        print(f"Video {name} already exists")
+        continue
+    
+    else: 
+        
+        writer = cv2.VideoWriter(name, cv2.VideoWriter_fourcc(*'mp4v'), 10.0, (frame_width, frame_height))
 
-        # Display the annotated frame
-        cv2.imshow("YOLOv11 eiders", annotated_frame)
-        writer.write(annotated_frame)
+        # Loop through the video frames
+        while cap.isOpened():
+            # Read a frame from the video
+            success, frame = cap.read()
 
-        # Break the loop if 'q' is pressed
-        if cv2.waitKey(1) & 0xFF == ord("q"):
-            break
-    else:
-        # Break the loop if the end of the video is reached
-        break
+            if success:
+                # Run YOLOv8 tracking on the frame, persisting tracks between frames
+                results = model.predict(frame, 
+                            stream=True, 
+                            save = True,
+                            show = False, 
+                            save_frames = False)
 
-# Release the video capture object and close the display window
-writer.release()
-cap.release()
-cv2.destroyAllWindows()
+                # Iterate over the results generator
+                for result in results:
+                    # Visualize the results on the frame
+                    annotated_frame = result.plot()
+
+                    # Write the annotated frame to the video writer
+                    writer.write(annotated_frame)
+
+            else:
+                # Break the loop if the end of the video is reached
+                break
+
+        # Release the video capture and writer objects, and close the display window
+        cap.release()
+        writer.release()
+        cv2.destroyAllWindows()
 
 
-  
-# Run example 
-# python3 code/yolov8/yolo_inference_multiple.py
+
+
